Split ACA infrastructure god class into multiple files
